fix bad imports (-types.ts importing from non -types.ts)
